db and multitenancy configuration details refactored
