+++ src/auth-service/app.js
@@ -10,7 +10,7 @@ const bodyParser = require("body-parser");
 const routes = require("@routes/index");
 const constants = require("@config/constants");
 const logger = log4js.getLogger(`${constants.ENVIRONMENT} -- app entry`);
-const mongodb = require("@config/dbConnection");
+const { mongodb } = require("@config/dbConnection");
 mongodb;
 
 const { logText } = require("@utils/log");
