+++ src/auth-service/models/Group.js
@@ -3,9 +3,9 @@ const ObjectId = mongoose.Schema.Types.ObjectId;
 const { Schema } = mongoose;
 const validator = require("validator");
 var uniqueValidator = require("mongoose-unique-validator");
-const { logObject, logElement, logText } = require("../utils/log");
+const { logObject, logElement, logText } = require("@utils/log");
 const isEmpty = require("is-empty");
-const { getModelByTenant } = require("../utils/multitenancy");
+const { getModelByTenant } = require("@config/dbConnection");
 const httpStatus = require("http-status");
 
 const GroupSchema = new Schema(
