+++ src/auth-service/package.json
@@ -1,91 +1,91 @@
 {
-    "name": "auth-service",
-    "version": "0.0.0",
-    "private": true,
-    "main": "./bin/www",
-    "scripts": {
-        "start": "./bin/www",
-        "stage": "NODE_ENV=staging node ./bin/www",
-        "dev-mac": "NODE_ENV=development nodemon ./bin/www",
-        "stage-mac": "NODE_ENV=staging nodemon ./bin/www",
-        "prod-mac": "NODE_ENV=production nodemon ./bin/www",
-        "dev-pc": "set NODE_ENV=development&&nodemon ./bin/www",
-        "stage-pc": "set NODE_ENV=staging&&nodemon ./bin/www",
-        "prod-pc": "set NODE_ENV=production&&nodemon ./bin/www",
-        "create-env": "printenv > .env",
-        "test": "mocha ./**/test/ut_*.js --exit"
-    },
-    "_moduleAliases": {
-        "@root": ".",
-        "@models": "models/",
-        "@utils": "utils",
-        "@controllers": "controllers",
-        "@config": "config",
-        "@middleware": "middleware",
-        "@scripts": "scripts",
-        "@routes": "routes"
-    },
-    "dependencies": {
-        "@google-cloud/datastore": "^5.1.0",
-        "@google-cloud/debug-agent": "^4.0.4",
-        "@log4js-node/slack": "^1.0.0",
-        "@mailchimp/mailchimp_marketing": "^3.0.75",
-        "app-module-path": "^2.2.0",
-        "bcrypt": "^5.0.0",
-        "body-parser": "^1.20.2",
-        "clean-deep": "^3.4.0",
-        "company-email-validator": "^1.0.6",
-        "compression": "^1.7.4",
-        "connect-mongo": "^3.2.0",
-        "continuation-local-storage": "^3.2.1",
-        "cookie-parser": "~1.4.3",
-        "crypto": "^1.0.1",
-        "debug": "~2.6.9",
-        "dotenv": "^8.2.0",
-        "email-existence": "^0.1.6",
-        "express": "^4.18.1",
-        "express-formidable": "^1.2.0",
-        "express-session": "^1.17.3",
-        "express-validation": "^1.0.3",
-        "express-validator": "^6.12.1",
-        "firebase-admin": "^11.5.0",
-        "firebase-functions": "^4.2.1",
-        "generate-password": "^1.5.1",
-        "gstore-node": "^7.2.4",
-        "helmet": "^3.21.1",
-        "http-status": "^1.4.0",
-        "is-empty": "^1.2.0",
-        "jade": "~1.11.0",
-        "joi": "^14.3.1",
-        "jsonrepair": "^3.0.2",
-        "jsonwebtoken": "^8.5.1",
-        "kafkajs": "^2.2.3",
-        "log4js": "^6.7.0",
-        "md5": "^2.3.0",
-        "module-alias": "^2.2.2",
-        "moment-timezone": "^0.5.41",
-        "mongoose": "^5.12.14",
-        "mongoose-unique-validator": "^2.0.3",
-        "morgan": "~1.9.0",
-        "nodemailer": "^6.6.1",
-        "passport": "^0.6.0",
-        "passport-auth-token": "^1.0.1",
-        "passport-jwt": "^4.0.0",
-        "passport-local": "^1.0.0",
-        "saslprep": "^1.0.3",
-        "serve-favicon": "~2.4.5",
-        "tslib": "^2.4.0",
-        "validator": "^11.1.0",
-        "winston": "^3.8.2",
-        "winston-mongodb": "^5.1.1",
-        "xoauth2": "^1.2.0"
-    },
-    "devDependencies": {
-        "chai": "^4.3.4",
-        "chai-http": "^4.3.0",
-        "faker": "^5.5.3",
-        "mocha": "^8.4.0",
-        "nodemon": "^1.19.3",
-        "sinon": "^11.1.1"
-    }
+  "name": "auth-service",
+  "version": "0.0.0",
+  "private": true,
+  "main": "./bin/www",
+  "scripts": {
+    "start": "./bin/www",
+    "stage": "NODE_ENV=staging node ./bin/www",
+    "dev-mac": "NODE_ENV=development nodemon --trace-warnings ./bin/www",
+    "stage-mac": "NODE_ENV=staging nodemon ./bin/www",
+    "prod-mac": "NODE_ENV=production nodemon ./bin/www",
+    "dev-pc": "set NODE_ENV=development&&nodemon ./bin/www",
+    "stage-pc": "set NODE_ENV=staging&&nodemon ./bin/www",
+    "prod-pc": "set NODE_ENV=production&&nodemon ./bin/www",
+    "create-env": "printenv > .env",
+    "test": "mocha ./**/test/ut_*.js --exit"
+  },
+  "_moduleAliases": {
+    "@root": ".",
+    "@models": "models/",
+    "@utils": "utils",
+    "@controllers": "controllers",
+    "@config": "config",
+    "@middleware": "middleware",
+    "@scripts": "scripts",
+    "@routes": "routes"
+  },
+  "dependencies": {
+    "@google-cloud/datastore": "^5.1.0",
+    "@google-cloud/debug-agent": "^4.0.4",
+    "@log4js-node/slack": "^1.0.0",
+    "@mailchimp/mailchimp_marketing": "^3.0.75",
+    "app-module-path": "^2.2.0",
+    "bcrypt": "^5.0.0",
+    "body-parser": "^1.20.2",
+    "clean-deep": "^3.4.0",
+    "company-email-validator": "^1.0.6",
+    "compression": "^1.7.4",
+    "connect-mongo": "^3.2.0",
+    "continuation-local-storage": "^3.2.1",
+    "cookie-parser": "~1.4.3",
+    "crypto": "^1.0.1",
+    "debug": "~2.6.9",
+    "dotenv": "^8.2.0",
+    "email-existence": "^0.1.6",
+    "express": "^4.18.1",
+    "express-formidable": "^1.2.0",
+    "express-session": "^1.17.3",
+    "express-validation": "^1.0.3",
+    "express-validator": "^6.12.1",
+    "firebase-admin": "^11.5.0",
+    "firebase-functions": "^4.2.1",
+    "generate-password": "^1.5.1",
+    "gstore-node": "^7.2.4",
+    "helmet": "^3.21.1",
+    "http-status": "^1.4.0",
+    "is-empty": "^1.2.0",
+    "jade": "~1.11.0",
+    "joi": "^14.3.1",
+    "jsonrepair": "^3.0.2",
+    "jsonwebtoken": "^8.5.1",
+    "kafkajs": "^2.2.3",
+    "log4js": "^6.7.0",
+    "md5": "^2.3.0",
+    "module-alias": "^2.2.2",
+    "moment-timezone": "^0.5.41",
+    "mongoose": "^5.12.14",
+    "mongoose-unique-validator": "^2.0.3",
+    "morgan": "~1.9.0",
+    "nodemailer": "^6.6.1",
+    "passport": "^0.6.0",
+    "passport-auth-token": "^1.0.1",
+    "passport-jwt": "^4.0.0",
+    "passport-local": "^1.0.0",
+    "saslprep": "^1.0.3",
+    "serve-favicon": "~2.4.5",
+    "tslib": "^2.4.0",
+    "validator": "^11.1.0",
+    "winston": "^3.8.2",
+    "winston-mongodb": "^5.1.1",
+    "xoauth2": "^1.2.0"
+  },
+  "devDependencies": {
+    "chai": "^4.3.4",
+    "chai-http": "^4.3.0",
+    "faker": "^5.5.3",
+    "mocha": "^8.4.0",
+    "nodemon": "^1.19.3",
+    "sinon": "^11.1.1"
+  }
 }
