+++ src/auth-service/utils/create-network.js
@@ -1,5 +1,5 @@
 const NetworkSchema = require("../models/Network");
-const { getModelByTenant } = require("./multitenancy");
+const { getModelByTenant } = require("@config/dbConnection");
 const { logElement, logText, logObject } = require("./log");
 const generateFilter = require("./generate-filter");
 const httpStatus = require("http-status");
