+++ src/auth-service/utils/test/ut_request.js
@@ -11,7 +11,7 @@ chai.use(chaiHttp);
 const UserSchema = require("../../models/User");
 const DefaultsSchema = require("../../models/Defaults");
 const requestUtil = require("../request");
-const { getModelByTenant } = require("../multitenancy");
+const { getModelByTenant } = require("@config/dbConnection");
 const { logObject, logElement, logText } = require("../log");
 
 const UserModel = (tenant) => {
